Improve oscillator period precision
Using an integer for the oscillator period was too imprecise. This
lead to some higher notes being audibly out of tune, especially in
chords. The solution is to use fixed-point arithmetic for phase. This
brings the detuning below 2 cents for most notes that matter.

+++ Src/modules/structures.h
@@ -11,7 +11,7 @@ typedef struct {
 } oscillator_settings;
 
 typedef struct {
-    uint16_t phase;
+    uint32_t phase;
 } oscillator_state;
 
 typedef enum {
